# public/api/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ownHackathon API Overview",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/token/refresh": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Refresh the access token",
                "description": "Generates a new short-lived access token using a valid refresh token. \n\nThis endpoint allows the client to maintain a session without requiring the user to re-enter their credentials. The refresh token must be provided via the configured security mechanism (e.g., Authorization Header or HTTP-only Cookie).",
                "operationId": "requestRegenerateAccessToken",
                "responses": {
                    "200": {
                        "description": "Successfully issued a new access token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessToken"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. This happens if the refresh token is expired, revoked, or invalid. The user must perform a full login again.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "refreshToken": []
                    }
                ]
            }
        },
        "/account/activation/{token}": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Finalize account registration and activation",
                "description": "Completes the registration process. This endpoint validates the activation token from the email and creates the actual user account with the provided username and password. \n\nNote: The account is only persisted in the database after this step is successful.",
                "operationId": "accountActivation",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "The secret activation token received via email.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The account credentials to be set for the new account.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountRegistration"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account successfully created and activated. The user can now log in."
                    },
                    "400": {
                        "description": "Validation failed. This can happen if:\n1. The token is invalid or has expired.\n2. The password does not meet the security requirements.\n3. The chosen account name is already taken.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/account/password/forgotten": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Request a token for password reset. Sending via E-Mail",
                "description": "Initiates the password recovery process for the given email address. \n\n**Security Note:** To prevent user enumeration, this endpoint always returns a 200 OK status. An attacker cannot determine whether an account exists for a specific email address by observing the API response.",
                "operationId": "requestPasswordReset",
                "requestBody": {
                    "description": "The email address associated with the account.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EMail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The request was accepted. If an account with the provided email address exists, an email containing a password reset link will be sent shortly."
                    }
                }
            }
        },
        "/account/password/{token}": {
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Reset account password using a token",
                "description": "Sets a new password for an account. This is the final step of the password recovery process. The request requires a valid, non-expired reset token that was previously sent to the user's email address.",
                "operationId": "resetPasswordWithToken",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "The secret reset token received via email.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The new password and its confirmation.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password has been successfully updated. The user can now log in with the new credentials."
                    },
                    "400": {
                        "description": "Request failed due to validation errors. Possible reasons:\n1. The reset token is invalid or has already been used.\n2. The token has expired.\n3. The new password does not meet the security requirements (e.g., too short).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/account": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Endpoint to register a new user account or request a password reset.",
                "description": "Starts the account process for the given email address. \n\n**Security Note:** This endpoint always returns a 200 OK status to prevent user enumeration. The user will not know if an account already exists based on the API response.",
                "operationId": "969568dfb3cf23c770ed1a16782476cd",
                "requestBody": {
                    "description": "The email address for the new account",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EMail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The request was processed successfully. Depending on the state of the account, one of the following actions will occur:\n1. **Account does not exist:** A token to activate the new account will be sent to the email address.\n2. **Account already exists:** A token to reset the password will be sent to the email address."
                    }
                }
            }
        },
        "/account/authentication": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Authenticate user and issue tokens",
                "description": "Authenticates a user using their credentials. On success, it returns a short-lived **AccessToken** (for API authorization) and a long-lived **RefreshToken** (to obtain new access tokens).",
                "operationId": "accountAuthentication",
                "requestBody": {
                    "description": "User credentials (email and password)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountAuthenticationData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful. The response contains both access and refresh tokens.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials. The email or password provided is incorrect.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied. The credentials are correct, but the account is currently restricted. Possible reasons: Account is locked, disabled, or the email address has not been verified yet.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AccountAuthenticationData": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The E-Mail from Account",
                        "type": "string"
                    },
                    "password": {
                        "description": "The Password from Account",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AccountPassword": {
                "properties": {
                    "password": {
                        "description": "The Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AccountRegistration": {
                "required": [
                    "accountName",
                    "password"
                ],
                "properties": {
                    "accountName": {
                        "description": "The Display Name for the Account",
                        "type": "string"
                    },
                    "password": {
                        "description": "The Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "EMail": {
                "properties": {
                    "email": {
                        "description": "The E-Mail",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AuthenticationResponse": {
                "properties": {
                    "accessToken": {
                        "description": "The access token after a valid log-in",
                        "type": "string"
                    },
                    "refreshToken": {
                        "description": "The refresh token after a valid log-in",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "HttpResponseMessage": {
                "properties": {
                    "statusCode": {
                        "description": "The Http Status Code",
                        "type": "integer",
                        "example": 400
                    },
                    "message": {
                        "description": "The Message",
                        "type": "string",
                        "example": "Bad request"
                    }
                },
                "type": "object"
            },
            "AccessToken": {
                "properties": {
                    "accessToken": {
                        "description": "The Token for authorized access",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "accessToken": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "refreshToken": {
                "type": "apiKey",
                "name": "Authentication",
                "in": "header"
            },
            "Client-Identification-String": {
                "type": "apiKey",
                "name": "x-ident",
                "in": "header"
            }
        }
    },
...
}